Revert the removal of support of NetBSD.
The previous release removed support for NetBSD in the traditional way
by dying with an `unsupported OS` message. Kamil Rytarowski, who works
on NetBSD, was kind enough to reach out and work on a fix for the
upstream libuv library: https://github.com/libuv/libuv/pull/2252 and to
give me a bit more information about why NetBSD isn't in the supported
platforms list.

Also, my previous decision to exclude NetBSD was based on evidence seen
up to this point on the CPAN Testers Matrix, but the versions of the OS
testing in most of those cases is kind of old. I'll look into more
testing on current NetBSD platforms before releasing in the future as I
don't want to prevent anyone from being able to use the library where it
might actually be useful.